[PC-8818] Move offerers listing to pages/Offerers/List/
