+++ src/components/pages/Offerers/List/OffererItem/OffererItem.jsx
@@ -4,9 +4,8 @@ import { Link } from 'react-router-dom'
 
 import Icon from 'components/layout/Icon'
 import { ReactComponent as AddOfferSvg } from 'icons/ico-plus.svg'
-
-import { pluralize } from '../../../../utils/pluralize'
-import { UNAVAILABLE_ERROR_PAGE } from '../../../../utils/routes'
+import { pluralize } from 'utils/pluralize'
+import { UNAVAILABLE_ERROR_PAGE } from 'utils/routes'
 
 const OffererItem = ({ offerer, physicalVenues, venues, isVenueCreationAvailable }) => {
   const { id, name, nOffers } = offerer || {}
