+++ src/components/pages/Offerers/List/OffererItem/__specs__/OffererItem.spec.jsx
