+++ src/components/pages/Offerers/List/OffererItem/__specs__/OffererItemContainer.spec.jsx
