+++ src/components/pages/Offerers/List/OffererItem/__specs__/PendingOffererItem.spec.jsx
